Fixed inheritance issue
