+++ src/Infrastructure/VolleyM.Infrastructure.AzureStorage/AzureTableConnection.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using LanguageExt;
 using Microsoft.Azure.Cosmos.Table;
@@ -94,53 +92,5 @@ namespace VolleyM.Infrastructure.AzureStorage
 
 		private static bool IsConflictError(StorageException e) =>
 			string.Compare("Conflict", e.Message, StringComparison.OrdinalIgnoreCase) == 0;
-
-		public EitherAsync<Error, Unit> ConfigureTables()
-		{
-			var conn = OpenConnection();
-
-			return conn.MapAsync(async client =>
-				{
-					var tables = GetTablesForContext();
-
-					var createTasks = Enumerable.Select<string, Task<bool>>(tables, table =>
-					{
-						var tableRef = client.GetTableReference(table);
-						return tableRef.CreateIfNotExistsAsync();
-					}).ToList();
-
-					await Task.WhenAll(createTasks);
-
-					return Unit.Default;
-				})
-				.ToAsync();
-		}
-
-		public EitherAsync<Error, Unit> CleanTables()
-		{
-			var conn = OpenConnection();
-
-			return conn.MapAsync(async client =>
-				{
-					var tables = GetTablesForContext();
-
-					var deleteTasks = Enumerable.Select<string, Task<bool>>(tables, table =>
-					{
-						var tableRef = client.GetTableReference(table);
-						return tableRef.DeleteIfExistsAsync();
-					});
-
-					await Task.WhenAll(deleteTasks);
-
-					return Unit.Default;
-				})
-				.ToAsync();
-		}
-
-		/// <summary>
-		/// When overriden should return all the tables particular configuration should be responsible for.
-		/// </summary>
-		/// <returns></returns>
-		protected abstract IEnumerable<string> GetTablesForContext();
 	}
 }
